Improvements to Sink Tool and general code
The following updates are to show how the the UI for a python toolbox tool can be improved, I have been in discussion with John about it. A test to see if I have understand how git hub functions I have improved only the Sink tool. The following edits were:
1] make input parameter a raster layer so it out be accessed from the map
2] Create a simple function to convert layer objects to full path which is required by WBT
3] Amended the item description so parameters now provide useful information when you click on them
4] Created a nice image that displays in tool.
5] Directly edited the WBT python module and put in a flag that hides the windows command line window that needless appears so tool behaves more like existing tools