Example improvements to PolygonsToLines tool
If you are not keen in having many XML files in the root folder (I have asked the ESRI community about that, we shall see what they say about that) then you need to make the tools smarter by validating the inputs. It is very easy to provide data that WBT cannot process, for example a file geodatabase featureclass. I don't think its good enough to say well it does say that in the manual.  The tools need to proactively ensure the correct inputs are going to WBT especially if you don't want to document them via the item description approach. It quickly becomes irritating and frustrating if one tries to input data that WBT throws an error at and their is no guidance other than delving into an online manual.

So I have worked up some sample code for the PolygonsToLines tool that could easily be embedding into your automation script that checks for 2 properties, it has a coordinate system and its a shapefile. This code is in the updateMessages function and by setting error warning you 1) stop the tool from trying to execute with bad data for WBT and 2) provide feed back as to why its wrong so the user can correct it, no need to go hunting in a manual.

This defensive approach makes the tools way more robust and the user experience much slicker!

For other tools, I've already blown up WBT by providing it rasters that were ESRI grids and rasters that did not have a coordinate system.  So at the minimum the tools need to be testing for data type and presence of coordinate systems.

Clearly tools that take numeric values need to know if negative numbers are acceptable.