Code tweaked to show how to get over schema changes
Ok this was way more problematic than I had anticipated. I think the main issue is that WBT edits the dataset output the ArcMap processing thread (if that is the correct terminology?) So ArcMap never knows about any changes. ArcMap also creates a file lock on the shapefile if the layer is loaded in the map. WBT does not seem to honour that so changes to a shapefile by WBT are not seen.

I managed to resolve it and example code is in the AddPointCoordinatesToTable tool. I've also had had import the `time` module to introduce a delay so that the OS can catch up. Essential the code kicks out the layer from the Map, runs WBT tool then loads the layer back in.

The limitations of this approach is that any styling is dropped but I'm sure this could be overcome by creating a temporary layerfile and load that back. If that is an issue I can work up the code for you?