Update WhiteboxTools.pyt
Updated the help links to the current site of the user manual on www.whiteboxgeo.com. I'm not sure if this file is maintained manually or if there is a script that automatically generates it though. If it is the later, then the script should be updated as well so as not to override this change.